changed some sendrecvs to be double

+++ main.cpp
@@ -224,8 +224,8 @@ int main(int argc, char* argv[]){
 				MPI_Sendrecv(&(localWeights[0]), particlesToReceive, MPI_INT, nextRank, 1, &(tempWeights[0]), particlesToReceive, MPI_INT, prevRank, 1, MPI_COMM_WORLD, &status);
 				MPI_Sendrecv(&(tempArray_s_x[0]),  particlesToReceive, MPI_INT, nextRank, 2, &(tempArray_s_x[0]), particlesToReceive, MPI_INT, prevRank, 2, MPI_COMM_WORLD, &status);
 				MPI_Sendrecv(&(tempArray_s_y[0]),  particlesToReceive, MPI_INT, nextRank, 3, &(tempArray_s_y[0]), particlesToReceive, MPI_INT, prevRank, 3, MPI_COMM_WORLD, &status);
-				MPI_Sendrecv(&(tempArray_f_x[0]),  particlesToReceive, MPI_INT, nextRank, 4, &(tempArray_f_x[0]), particlesToReceive, MPI_INT, prevRank, 2, MPI_COMM_WORLD, &status);
-				MPI_Sendrecv(&(tempArray_f_y[0]),  particlesToReceive, MPI_INT, nextRank, 5, &(tempArray_f_y[0]), particlesToReceive, MPI_INT, prevRank, 3, MPI_COMM_WORLD, &status);
+				MPI_Sendrecv(&(tempArray_f_x[0]),  particlesToReceive, MPI_DOUBLE, nextRank, 4, &(tempArray_f_x[0]), particlesToReceive, MPI_DOUBLE, prevRank, 2, MPI_COMM_WORLD, &status);
+				MPI_Sendrecv(&(tempArray_f_y[0]),  particlesToReceive, MPI_DOUBLE, nextRank, 5, &(tempArray_f_y[0]), particlesToReceive, MPI_DOUBLE, prevRank, 3, MPI_COMM_WORLD, &status);
 				MPI_Sendrecv(&(pointerForLocalArray[0]),  particlesToReceive, MPI_INT, nextRank, 6, &(pointerForTempArray[0]), particlesToReceive, MPI_INT, prevRank, 4, MPI_COMM_WORLD, &status);
 				/*j = 0;
 				for(i = 0; i < particlesToReceive; i++){
@@ -364,8 +364,8 @@ int main(int argc, char* argv[]){
 				MPI_Sendrecv(&(localWeights[0]), particlesToReceive, MPI_INT, nextRank, 1, &(tempWeights[0]), particlesToReceive, MPI_INT, prevRank, 1, MPI_COMM_WORLD, &status);
 				MPI_Sendrecv(&(tempArray_s_x[0]),  particlesToReceive, MPI_INT, nextRank, 2, &(tempArray_s_x[0]), particlesToReceive, MPI_INT, prevRank, 2, MPI_COMM_WORLD, &status);
 				MPI_Sendrecv(&(tempArray_s_y[0]),  particlesToReceive, MPI_INT, nextRank, 3, &(tempArray_s_y[0]), particlesToReceive, MPI_INT, prevRank, 3, MPI_COMM_WORLD, &status);
-				MPI_Sendrecv(&(tempArray_f_x[0]),  particlesToReceive, MPI_INT, nextRank, 4, &(tempArray_f_x[0]), particlesToReceive, MPI_INT, prevRank, 2, MPI_COMM_WORLD, &status);
-				MPI_Sendrecv(&(tempArray_f_y[0]),  particlesToReceive, MPI_INT, nextRank, 5, &(tempArray_f_y[0]), particlesToReceive, MPI_INT, prevRank, 3, MPI_COMM_WORLD, &status);
+				MPI_Sendrecv(&(tempArray_f_x[0]),  particlesToReceive, MPI_DOUBLE, nextRank, 4, &(tempArray_f_x[0]), particlesToReceive, MPI_DOUBLE, prevRank, 2, MPI_COMM_WORLD, &status);
+				MPI_Sendrecv(&(tempArray_f_y[0]),  particlesToReceive, MPI_DOUBLE, nextRank, 5, &(tempArray_f_y[0]), particlesToReceive, MPI_DOUBLE, prevRank, 3, MPI_COMM_WORLD, &status);
 				MPI_Sendrecv(&(pointerForLocalArray[0]),  particlesToReceive, MPI_INT, nextRank, 6, &(pointerForTempArray[0]), particlesToReceive, MPI_INT, prevRank, 4, MPI_COMM_WORLD, &status);
 				
 				//Calculate forces
